Move framework specific code to submodules
